Start InputText refactor
