+++ packages/components/src/Form/Fields/Field.tsx
@@ -31,7 +31,7 @@ import omit from 'lodash/omit'
 import pick from 'lodash/pick'
 import { Paragraph, Text } from '../../Text'
 import { FieldsetContext } from '../Fieldset'
-import { inputHeight } from '../Inputs/InputText/InputText'
+import { inputHeight } from '../Inputs/InputText'
 import { Label } from '../Label'
 import { VisuallyHidden } from '../../VisuallyHidden'
 import { ValidationMessage } from '../ValidationMessage'
