+++ packages/components/src/Form/Inputs/InputText/index.ts
@@ -25,3 +25,4 @@
  */
 
 export * from './InputText'
+export * from './InputTextBase'
